edits to Game1.cs
