+++ Game Project/Game1.cs
@@ -8,19 +8,14 @@ namespace Game_Project
     {
         private GraphicsDeviceManager _graphics;
         private SpriteBatch _spriteBatch;
-        private SpriteFont font;
 
         // Interfaces to use
         public IController keyboard;
-        public IController mouse;
-        private ISprite sprite;
-
-        // Necessary so that command recievers can display things correctly. These would be stored somewhere different and would be
-        // loaded in a different way, this is by no means a final or best design
-        public Texture2D PlayerTexture { get; private set; }
-        public Vector2 Location { get; private set; }
-        public int rows;
-        public int columns;
+        public Player player;
+        public BatEnemy batEnemy;
+        public TileManager tiles;
+        public EnemyManager enemies;
+        public ItemManager items;
 
         public Game1()
         {
@@ -38,12 +33,13 @@ namespace Game_Project
         protected override void Initialize()
         {
             //GraphicsDevice.SamplerStates[0] = SamplerState.PointClamp;
-            Location = new Vector2(GraphicsDevice.Viewport.Width / 2 - 48, GraphicsDevice.Viewport.Height / 2 - 64);
-            rows = 3;
-            columns = 4;
 
             keyboard = new KeyboardController();
-            mouse = new MouseController();
+            player = new Player(spriteBatch);
+            batEnemy = new BatEnemy();
+            tiles = new TileManager();
+            enemies = new EnemyManager();
+            items = new ItemManager();
 
             base.Initialize();
         }
@@ -57,11 +53,6 @@ namespace Game_Project
             _spriteBatch = new SpriteBatch(GraphicsDevice);
 
             keyboard.LoadContent(this);
-            mouse.LoadContent(this);
-
-            PlayerTexture = Content.Load<Texture2D>("HyperLightWalkingRight");
-            font = Content.Load<SpriteFont>("Text");
-            // Set default sprite state
         }
 
         /// <summary>
@@ -74,9 +65,9 @@ namespace Game_Project
             if (GamePad.GetState(PlayerIndex.One).Buttons.Back == ButtonState.Pressed || Keyboard.GetState().IsKeyDown(Keys.Escape))
                 Exit();
 
-            sprite.Update();
+            player.Update();
             keyboard.Update();
-            mouse.Update();
+            batEnemy.bat.Update();
 
             base.Update(gameTime);
         }
@@ -89,20 +80,7 @@ namespace Game_Project
         {
             GraphicsDevice.Clear(Color.CornflowerBlue);
 
-            // Draw the text
-            _spriteBatch.Begin();
-            _spriteBatch.DrawString(font, "Credits\nProgram made by: Sebastian King\nhttps://www.pinterest.com/pin/789326272163794175/", new Vector2(125, 400), Color.Black);
-            _spriteBatch.End();
-
-            // Draw the sprite
-            sprite.Draw(_spriteBatch, new Vector2(5f));
-
             base.Draw(gameTime);
-        }
-
-        public void setSprite(ISprite sprite)
-        {
-            this.sprite = sprite;
         }
     }
 }
